requirements.txt
Flask
schedule
weather-api